Since the change is so small now, make Image async again
We were going to have two image classes Image and AsyncImage
but since the "sync path" has changed so little from the original
I think it's safe to just one Image.py which can be used in
sync or async mode. The async path functionally should be
basically unchanged from before the async changes.
